Add user prompt and computer choice, comment edits
In this commit, I have added a prompt for the user to type in rock,
paper, or scissors. I have also added a function for the computer to
generate a random integer between 1 and 3, with comments added to the
process I laid out in the previous commit to document later need to
convert playerSelection to an integer. I may edit the prompt later,
after further research, to create a button option rather than a text
prompt.