Add convert function for computerSelection
In this commit, I have added an if/elseif/else function to convert the
integer value to the string value (rock, paper, or scissors) in order to
make the computer compete with the user properly. I could have done this
instead by converting the user's selection to an integer and adjust the
process accordingly to compare the two integers, rather than two
strings, but I chose this route as it feels more efficient to only enact
one conversion rather than three (as I would eventually convert both
user and computer values to strings). I adjusted the comments to
properly document the process and inform the direction of the process
based on the comparison route I chose.