Fix errors, make playerSelection case-insensitive
I reworked the prompt variable with a .toLowerCase function in order to
ensure that the value of playerSelection would always be lowercase to
avoid errors in comparison later on. I also fixed errors I noticed in
the parameters of the roundOutcome usage of the playRound function, and
my use of 'let' vs. 'var', so the roundOutcome variable returns as it
should.