Reorder functions for correct output
After discovering that the game() function is actually first invoked in
the gameLoop() function, I recognized that the declareWinner() function
would not be operational before the invocation of the gameLoop()
function. I have moved the declareWinner() function to follow the
invocation of the gameLoop() function, resulting in fully operational
code.