Add buttons for playerSelection
In this commit, I have added button elements to the DOM and added a
click-based event listener to each, which allows the buttons to be used
instead of a prompt to determine playerSelection and run the playRound()
function. The game is not functioning properly yet, as I am refactoring
code to make the GUI functional, but it's a start.