New JavaScript file, updates to button events
In this commit, I moved over to a new JavaScript file to eliminate
confusion as I continue. The script source file in the HTML has been
updated accordingly.

I have also updated the button functions to include getting
computerChoice and computerSelection, as I believed that was the cause
of one of the functional issues I have found in trying to test the
button usage version of gameplay with the console. I am still having
trouble getting this to function the way I want it to in the console,
but I think the functionality will improve as I continue to explore the
implementation of and transition to the UI.