Added divs, restart button
In this commit, I removed unused lines from the javascript file and
primarily added divs to the HTML. I added a div for the playerButtons,
and moved the playerButtons class to the div containing all buttons
within this class, instead of having the class declared on each button.
I also added a div where the results will be displayed, and a div with a
restart button for greater continued gameplay once the project is
complete.