Added gameStatus, minor class to id switch
In this update, I added a function which makes the first addition to the
container for the results of the game. The gameStatus() function (name
may be changed later on) currently displays the message stating which
weapon the player chose. This function is called and the message
variable's value declared within each button's respective function.
I also revised the results container's class to an id, for easy handling
later on.