Update playRound; gameStatus & playerChoice edits
In this commit, I updated the playRound function in a few ways. First, I
reverted the function back to multiple if/else if statements for each
standoff (except for the tie, which was an inclusive statement) so as to
display custom messages based on the involved selections for each
standoff. I did this so as to include the wording of each rule which is
particularly memorable in Rock Paper Scissors Lizard Spock.

Additonally, I renamed the original gameStatus() function to
playerChoice(), and renamed the variable therein to message1, rather
than just message. Then, I added a new gameStatus() function for the
purpose of displaying the round outcome message, using the variable
message2. I prefer to number the message variables this way because they
are both messages which will be appended to the results container, and
both fall under the distinction of results messages.