Add ending() & gameOver() functions
In this commit, I've added the ending() function to add a message to end
the game once either the player or the computer has accrued 5 points,
stored in variable message3. I have also added the gameOver() function,
which, within the playRound function, tells the browser when to display
message3 (or rather, when the game ends).