Add scoreboard
In this commit, I added a scoreboard layout to the HTML document, with
a div and id for each piece of scoreboard content, including the
player's score, the computer's score, and the current round number. I
then created a constant in the JS document for each scoreeboard div id.
Within the playRound() function, I then added an innerHTML property to
each constant to make each scoreboard div display the values of
currentRound, playerPoints, and computerPoints (the respective
variables for each piece of content) as text in their relevant divs (and
to update when changed).