Add victory banner and endgame provisions
In this commit, I have added a victory banner to the scoreboard in the
HTML doc. From there, within the playRound() function, I tried a new
approach at ending the game using the victory banner in an if/else if
structure which would determine when to display the victory banner
(based on player or computer accruing 5 points) and what text content
to display depending on which party accrued 5 points first. Currently,
the provisions I have made for endgame are not forcing the user to reset
the game following a win or loss (the buttons did not disable the way
I intended, and the playRound() function still operates when buttons
are pressed).