Reattempts at gameOver()/forcing endgame reset
In this commit, I took a new approach to forcing the user to reset the
game after a win or loss, but was unsuccessful. I removed the original
gameOver() function (and may revert to bring it back later), and created
a new one wihin the if/else if structure from the previous commit. The
new gameOver() function was created to remove the event listeners from
the buttons in attempt to get them to stop calling the playRound()
function. Will investigate further.