Force endgame reset using disabled property
In this commit, I removed the gameOver() function and set the disabled
property to true on each individual button's constant within the
if/else if structure I've set for the endgame. Now, the buttons will
grey out and cease to function once either the player or the computer
has accrued 5 points.