UI Styling: Fonts
In this commit, I made edits to update the HTML and style the UI using
HTML and CSS edits. I added a meta tag to the HTML header for mobile
rendering, and added a link to a new stylesheet. From there, I added
links in the HTML header (marked by comments) to the fonts I chose for
the header and body--custom fonts hosted by Google. I then added CSS
properties to determine the fonts based on the header size.