UI Styling: Imagery & Fixes
In this commit, I made revisions to the HTML doc and additions to the
stylesheet to continue customizing the visual layout and appeal of the
UI. I added an images folder with images I will eventually use for the
buttons, as well as an image for the background. Then, in the HTML doc,
I switched the order to have the endgame class before the results class,
and I added a new div surrounding all content except the results div, so
as to move the results (with an additional h2 marking this section as a
game log) to the side, rather than part of one long display, which would
require the player to use a scroll bar. I moved the victory message from
outside the gameStatus div in the main-body div, and added padding and
flexbox properties to the victory message and the scoreboard markers so
as to center them appropriately. I also adjusted the font colors to make
the page, with the new background gradient, WCAG-compliant and easy to
read.