UI Styling: Buttons (hover, disabled)
In this commit, I added and styled hover and disabled states for the
player buttons, the hover state also applying to the restart button. In
attempt to stop an issue with the game log section bouncing when
hovering on the buttons, I added a margin to the playerButtons class
(which contains the buttons). I also added a cursor property to the
buttons to further prompt and support interactivity with the buttons.